dom/tests/mochitest/ajax/offline/test_identicalManifest.html

<html xmlns="http://www.w3.org/1999/xhtml" manifest="http://mochi.test:8888/tests/dom/tests/mochitest/ajax/offline/simpleManifest.cacheManifest">
<head>
<title>identical manifest test</title>

<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="/tests/dom/tests/mochitest/ajax/offline/offlineTests.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

var gGotChecking = false;
var gGotDownloading = false;

function finishTest()
{
  OfflineTest.teardownAndFinish();
}

function noUpdate()
{
  OfflineTest.ok(gGotChecking, "Should get a checking event");
  OfflineTest.ok(!gGotDownloading, "Should not get a downloading event");

  var entries = [
    // The document that requested the manifest should be in the cache
    [window.location.href, true],

    ["http://mochi.test:8888/tests/SimpleTest/SimpleTest.js", true],
    ["http://mochi.test:8888/tests/dom/tests/mochitest/ajax/offline/offlineTests.js", true]
  ];
  OfflineTest.checkCacheEntries(entries, finishTest);
}

function manifestUpdated()
{
  OfflineTest.ok(gGotChecking, "Should get a checking event");
  OfflineTest.ok(gGotDownloading, "Should get a downloading event");

  var entries = [
    // The manifest itself should be in the cache
    ["http://mochi.test:8888/tests/dom/tests/mochitest/ajax/offline/simpleManifest.cacheManifest", true],

    // The document that requested the manifest should be in the cache
    [window.location.href, true],

    // The entries from the manifest should be in the cache
    ["http://mochi.test:8888/tests/SimpleTest/SimpleTest.js", true],
    ["http://mochi.test:8888/tests/dom/tests/mochitest/ajax/offline/offlineTests.js", true],

    // The bad entries from the manifest should not be in the cache
    ["bad:/uri/invalid", false]
  ];
  OfflineTest.checkCacheEntries(entries, manifestUpdatedContinue);
}

function manifestUpdatedContinue()
{
  // Now make sure applicationCache.update() does what we expect.
  applicationCache.oncached = OfflineTest.failEvent;
  applicationCache.onnoupdate = OfflineTest.priv(noUpdate);
  
  gGotChecking = false;
  gGotDownloading = false;
  applicationCache.update();
}

if (OfflineTest.setup()) {
  applicationCache.onerror = OfflineTest.failEvent;
  applicationCache.onnoupdate = OfflineTest.failEvent;
  applicationCache.onudpateready = OfflineTest.failEvent;

  applicationCache.onchecking = function() { gGotChecking = true; };
  applicationCache.ondownloading = function() { gGotDownloading = true; };
  applicationCache.oncached = OfflineTest.priv(manifestUpdated);
}

SimpleTest.waitForExplicitFinish();

</script>

</head>

<body>

</body>
</html>
